Feat: 코드 확정

See merge request s11-webmobile1-sub2/S11P12A501!188

+++ BackEnd/globalcc/src/main/java/com/ssafy/globalcc/domain/meetingSTT/dto/request/MeetingSTTSegment.java
@@ -8,7 +8,7 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MeetingSTTSegment {
-    private float start;
-    private float end;
+    private double start;
+    private double end;
     private String text;
 }
